+++ CHANGELOG.rst
@@ -14,6 +14,8 @@ Version 0.1.2 (Current Develop)
 * Update SciKit Learn to 0.22.2, Update NumPy to 1.18.3, and other dependencies.
 * Use pipelines for Random Forest and enable categorical variables
 * Fix EPW Day of week calculation. Add Day of Week (%A) and Day of Week Int (%w).
+* Add seaborn-qqplot to requirements
+* Fix bug passing downsampling value to validation (users needs to ensure there is a downsample key in the metamodel.json file)
 
 Version 0.1.1
 =============